requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.2
django-crispy-forms==2.4
pillow==11.2.1
sqlparse==0.5.3
